backend/requirements.txt
Django>=4.2,<5.0
celery>=5.3,<6.0
redis>=4.5
django-cors-headers>=4.0
gunicorn>=21.0
djangorestframework
djangorestframework-simplejwt
Pillow==10.2.0

# Optional but common
python-dotenv>=1.0
gunicorn>=21.0
psycopg2-binary>=2.9
celery>=5.3
redis>=4.5
drf-yasg==1.21.7